refactor(viewer): flatten natural_query files into lib/ directory
The natural_query/ subdirectory was not being served by the server's
include_dir!() macro which only copies top-level contents. Move all
files to flat nq_*.js files in src/viewer/assets/lib/ and update
import paths accordingly.
