--- site/viewer/lib/natural_query
@@ -1,1 +1,0 @@
-../../../src/viewer/assets/lib/natural_query
