+++ src/viewer/assets/lib/nq_prompt.js
